# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NexusPay API
  description: |
    A payment processing API that enables users to register, manage wallets, and transfer funds.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Rate Limiting
    Public endpoints are rate-limited to 15 requests per minute per IP address.
  version: 1.0.0
  contact:
    name: API Support
    email: support@nexuspay.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: localhost:3000
    description: Development server
  - url: api.nexuspay.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User registration and login endpoints
  - name: Users
    description: User management operations
  - name: Webhooks
    description: Webhook handlers for external services
  - name: Wallet
    description: Wallet management operations

paths:
  /healthx:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: |
        Returns the current health status of the service.
        This endpoint is public and rate-limited.
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: good for now

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Creates a new user account with email and password.
        Returns JWT access and refresh tokens upon successful registration.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: john.doe@example.com
              password: SecurePassword123!
              full_name: John Doe
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                email: john.doe@example.com
                full_name: John Doe
                jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        "400":
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Invalid input
                message: Email is required and must be valid
                status_code: 400
        "409":
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: User already exists
                message: A user with this email already exists
                status_code: 409

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticates a user with email and password.
        Returns JWT access and refresh tokens upon successful authentication.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: john.doe@example.com
              password: SecurePassword123!
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                email: john.doe@example.com
                full_name: John Doe
                jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        "401":
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid credentials
                message: The provided credentials are incorrect
                status_code: 401
        "404":
          description: Not found - user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: user not found
                message: No user found with this email
                status_code: 404

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refreshes the JWT access token using a valid refresh token.
        Use this endpoint to obtain a new access token when the current one expires.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
              example:
                jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: bmV3IHJlZnJlc2ggdG9rZW4...
        "404":
          description: Not found - invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: User not found
                message: Invalid or expired refresh token
                status_code: 404

  /users/test:
    get:
      tags:
        - Authentication
      summary: Test authentication
      description: |
        Protected endpoint to verify JWT token validity.
        Returns the claims encoded in the JWT token.
        Requires a valid JWT token in the Authorization header.
      operationId: testAuth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authentication successful, returns token claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenClaims"
              example:
                id: 123e4567-e89b-12d3-a456-426614174000
                exp: 1705317600
                iat: 1704712800
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401

  /users/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: |
        Logs out the current user by invalidating their session.
        Requires a valid JWT token in the Authorization header.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401

  /webhook/stripe:
    post:
      tags:
        - Webhooks
      summary: Stripe webhook
      description: |
        Handles incoming Stripe webhook events (payment_intent.succeeded, payment_intent.payment_failed, payment_intent.canceled).
        Validates the Stripe signature to ensure the request is from Stripe.
        This endpoint is public but signature-verified.
      operationId: stripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Stripe event ID
                  example: evt_test_webhook
                type:
                  type: string
                  description: Stripe event type
                  example: payment_intent.succeeded
                data:
                  type: object
                  description: Event data payload
      headers:
        - name: Stripe-Signature
          schema:
            type: string
          description: Stripe webhook signature for verification
      responses:
        "200":
          description: Webhook handled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
        "400":
          description: Invalid Stripe signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid signature
                message: Webhook signature verification failed
                status_code: 400

  /wallet/:
    get:
      tags:
        - Wallet
      summary: Get wallet
      description: |
        Retrieves the wallet details for the authenticated user.
        Returns the wallet ID, balance, and creation timestamp.
        Requires a valid JWT token in the Authorization header.
      operationId: getWallet
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Wallet
      summary: Top up wallet
      description: |
        Creates a Stripe payment intent to top up the wallet balance.
        Returns a client secret for the Stripe payment sheet.
        Requires a valid JWT token in the Authorization header.
      operationId: topUpWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  description: Amount to top up in smallest currency unit (e.g., cents)
                  example: 5000
      responses:
        "200":
          description: Payment intent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_secret:
                    type: string
                    description: Stripe payment intent client secret
                    example: pi_test123_secret_abc
                  amount:
                    type: integer
                    example: 5000
                  currency:
                    type: string
                    example: usd
        "400":
          description: Bad request - invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/:
    get:
      tags:
        - Users
      summary: Search users
      description: |
        Searches for users by name. Returns a list of matching users.
        Requires a valid JWT token in the Authorization header.
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          description: Name to search for
          required: false
          schema:
            type: string
            example: John
      responses:
        "200":
          description: Users found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token received from `/auth/register` or `/auth/login`.
        Include in the Authorization header: `Bearer <token>`

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          description: User's email address (must be valid and unique)
          example: john.doe@example.com
        password:
          type: string
          format: password
          description: |
            User's password. Must meet security requirements.
            Maximum length: 72 bytes (for bcrypt compatibility)
          example: SecurePassword123!
        full_name:
          type: string
          description: |
            User's full name. Must be at least 3 characters.
          minLength: 3
          example: John Doe

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's registered email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          description: User's password
          example: SecurePassword123!

    RegisterResponse:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        full_name:
          type: string
          description: User's full name
          example: John Doe
        jwt_token:
          type: string
          description: |
            JWT access token for authenticated requests.
            Use this in the Authorization header.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refresh_token:
          type: string
          description: |
            Refresh token for obtaining new access tokens.
            Store securely and never expose to client-side code.
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gdGhhdCBzaG91bGQgYmUga2VwdCBzZWNyZXRseQ==

    LoginResponse:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        full_name:
          type: string
          description: User's full name
          example: John Doe
        jwt_token:
          type: string
          description: JWT access token for authenticated requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gdGhhdCBzaG91bGQgYmUga2VwdCBzZWNyZXRseQ==

    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: |
            Valid refresh token received during login or registration.
            Must not be expired.
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4gdGhhdCBzaG91bGQgYmUga2VwdCBzZWNyZXRseQ==

    RefreshResponse:
      type: object
      properties:
        jwt_token:
          type: string
          description: New JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refresh_token:
          type: string
          description: New refresh token (may be same or new)
          example: bmV3IHJlZnJlc2ggdG9rZW4gdGhhdCBzaG91bGQgYmUga2VwdCBzZWNyZXRseQ==

    TokenClaims:
      type: object
      properties:
        id:
          type: string
          description: User ID from the token
          example: 123e4567-e89b-12d3-a456-426614174000
        exp:
          type: integer
          description: Token expiration timestamp (Unix epoch)
          example: 1705317600
        iat:
          type: integer
          description: Token issued at timestamp (Unix epoch)
          example: 1704712800

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Short error identifier
          example: Invalid input
        message:
          type: string
          description: Human-readable error message
          example: Email is required and must be valid
        status_code:
          type: integer
          description: HTTP status code
          example: 400

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        full_name:
          type: string
          description: User's full name
          example: John Doe
        created_at:
          type: string
          format: date-time
          description: Timestamp when user was created
          example: 2024-01-15T10:30:00Z

    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique wallet identifier
          example: 123e4567-e89b-12d3-a456-426614174001
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this wallet
          example: 123e4567-e89b-12d3-a456-426614174000
        balance:
          type: integer
          description: |
            Current wallet balance in smallest currency unit (e.g., cents).
            Divide by 100 to get the actual amount.
          example: 100000
        created_at:
          type: string
          format: date-time
          description: Timestamp when wallet was created
          example: 2024-01-15T10:30:00Z

    Transfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique transfer identifier
          example: 123e4567-e89b-12d3-a456-426614174002
        from_wallet_id:
          type: string
          format: uuid
          description: Source wallet ID
          example: 123e4567-e89b-12d3-a456-426614174001
        to_wallet_id:
          type: string
          format: uuid
          description: Destination wallet ID
          example: 123e4567-e89b-12d3-a456-426614174003
        amount:
          type: integer
          description: Transfer amount in smallest currency unit
          example: 5000
        status:
          type: string
          enum:
            - pending
            - completed
            - failed
          description: Current status of the transfer
          example: completed
        note:
          type: string
          description: Optional note about the transfer
          nullable: true
          example: Lunch money
        created_at:
          type: string
          format: date-time
          description: Timestamp when transfer was created
          example: 2024-01-15T10:30:00Z

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique transaction identifier
          example: 123e4567-e89b-12d3-a456-426614174004
        wallet_id:
          type: string
          format: uuid
          description: Wallet associated with this transaction
          example: 123e4567-e89b-12d3-a456-426614174001
        amount:
          type: integer
          description: Transaction amount in smallest currency unit
          example: 5000
        type:
          type: string
          enum:
            - debit
            - credit
          description: Transaction type (money out or in)
          example: debit
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
            - reversed
          description: Current status of the transaction
          example: completed
        transfer_id:
          type: string
          format: uuid
          description: Associated transfer ID (if applicable)
          nullable: true
          example: 123e4567-e89b-12d3-a456-426614174002
        created_at:
          type: string
          format: date-time
          description: Timestamp when transaction was created
          example: 2024-01-15T10:30:00Z

    ScheduledTransfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique scheduled transfer identifier
          example: 123e4567-e89b-12d3-a456-426614174005
        transfer_id:
          type: string
          format: uuid
          description: Associated transfer ID
          example: 123e4567-e89b-12d3-a456-426614174002
        scheduled_at:
          type: string
          format: date-time
          description: When the transfer is scheduled to execute
          example: 2024-01-20T10:00:00Z
        executed_at:
          type: string
          format: date-time
          description: When the transfer was actually executed
          nullable: true
          example: 2024-01-20T10:00:05Z
        created_at:
          type: string
          format: date-time
          description: Timestamp when scheduled transfer was created
          example: 2024-01-15T10:30:00Z

  responses:
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
            message: Missing or invalid authorization token
            status_code: 401

    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: internal_error
            message: An unexpected error occurred
            status_code: 500

security:
  - bearerAuth: []